+++ src/main/java/org/apiphany/lang/gzip/GZip.java
@@ -11,6 +11,7 @@ import java.util.zip.GZIPOutputStream;
 
 import org.apiphany.io.IOStreams;
 import org.apiphany.lang.Strings;
+import org.morphix.lang.JavaObjects;
 import org.morphix.reflection.Constructors;
 
 /**
@@ -86,7 +87,7 @@ public class GZip {
 	 * @return de-compressed string
 	 * @throws IOException on error
 	 */
-	public static String decompress(final byte[] body) throws IOException {
+	public static String decompressToString(final byte[] body) throws IOException {
 		try (GZIPInputStream gzipInputStream = new GZIPInputStream(new ByteArrayInputStream(body))) {
 			return Strings.toString(gzipInputStream, Strings.DEFAULT_CHARSET, IOStreams.DEFAULT_BUFFER_SIZE);
 		}
@@ -99,10 +100,42 @@ public class GZip {
 	 * @return de-compressed string
 	 * @throws IOException on error
 	 */
-	public static String decompress(final InputStream inputStream) throws IOException {
+	public static String decompressToString(final InputStream inputStream) throws IOException {
 		try (GZIPInputStream gzipInputStream = new GZIPInputStream(inputStream)) {
 			return Strings.toString(gzipInputStream, Strings.DEFAULT_CHARSET, IOStreams.DEFAULT_BUFFER_SIZE);
 		}
+	}
+
+	/**
+	 * De-compress a GZIP-ed byte array to a byte array.
+	 *
+	 * @param body GZIP byte array to decompress
+	 * @return de-compressed byte array
+	 * @throws IOException on error
+	 */
+	public static byte[] decompressToBytes(final byte[] body) throws IOException {
+		try (GZIPInputStream gis = new GZIPInputStream(new ByteArrayInputStream(body))) {
+			return IOStreams.toByteArray(gis);
+		}
+	}
+
+	/**
+	 * De-compress a GZIP-ed input (byte array or input stream) to a the same output type as the input.
+	 *
+	 * @param <T> input type
+	 *
+	 * @param input input to de-compress
+	 * @return de-compressed value in the same type as the input
+	 * @throws IOException on error
+	 */
+	public static <T> T decompress(final T input) throws IOException {
+		Object result = switch (input) {
+			case byte[] bytes -> decompressToBytes(bytes);
+			case InputStream is -> inputStream(is);
+			default -> throw new IllegalArgumentException("Cannot decompress object of type: " + input.getClass()
+					+ ", input must be byte[] or " + InputStream.class);
+		};
+		return JavaObjects.cast(result);
 	}
 
 	/**